week_1/requirements.txt
groq
python-dotenv
streamlit
openai